names correction
