+++ src/main/java/org/example/dto/TraineeRequest.java
@@ -1,0 +1,69 @@
+package org.example.dto;
+
+import java.time.LocalDate;
+
+public class TraineeRequest {
+    private Long id;
+    private String firstName;
+    private String lastName;
+    private Boolean isActive;
+    private LocalDate dateOfBirth;
+    private String address;
+
+    public TraineeRequest(Long id, String firstName, String lastName, Boolean isActive, LocalDate dateOfBirth, String address) {
+        this.id = id;
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.isActive = isActive;
+        this.dateOfBirth = dateOfBirth;
+        this.address = address;
+    }
+
+    public TraineeRequest(String firstName, String lastName, Boolean isActive, LocalDate dateOfBirth, String address) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.isActive = isActive;
+        this.dateOfBirth = dateOfBirth;
+        this.address = address;
+    }
+
+    public String getFirstName() {
+        return firstName;
+    }
+
+    public void setFirstName(String firstName) {
+        this.firstName = firstName;
+    }
+
+    public String getLastName() {
+        return lastName;
+    }
+
+    public void setLastName(String lastName) {
+        this.lastName = lastName;
+    }
+
+    public Boolean getActive() {
+        return isActive;
+    }
+
+    public void setActive(Boolean active) {
+        isActive = active;
+    }
+
+    public LocalDate getDateOfBirth() {
+        return dateOfBirth;
+    }
+
+    public void setDateOfBirth(LocalDate dateOfBirth) {
+        this.dateOfBirth = dateOfBirth;
+    }
+
+    public String getAddress() {
+        return address;
+    }
+
+    public void setAddress(String address) {
+        this.address = address;
+    }
+}
